Instead of opening tuntox buffer, use messages
The *Tuntox Proxy* buffer is mostly useful to those who know about the
low-level operation of crdt.el. So, instead of always opening that
process buffer when it is created, leave the process buffer in the
background and use process filters to post more user-friendly
notifications about tuntox connection status to *Messages*.

I'm not that familiar with process filters, but I saw that overriding
the default filter causes the buffer to no longer get updated by
default, so I copied the way the crdt-connect tuntox filter writes to
the process buffer into crdt-new-session as well.

Co-authored-by: Arte Ebrahimi <arteebrahimi@gmail.com>